Initial setup of Hi gem
Hi utilizes the 'colorize' gem to highlight specified words from
standard input. This is the initial version which does a simple find and
replace on STDIN using ARGV. It will always replace the word with a bold
red word, and it does not prevent multiple replacements.

This project was setup following this blog post:
https://hspazio.github.io/2017/build-a-cli-as-ruby-gem/
https://github.com/hspazio/weather